Moves state to app compnent
