+++ clients/react-movies/src/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Route, Switch } from 'react-router-dom'
 
 import './App.css';
@@ -6,6 +6,18 @@ import MovieList from './pages/MovieList';
 import MovieDetail from './pages/MovieDetail';
 
 function App() {
+  const [state, setState] = useState({ movies: [] });
+
+  useEffect(() => {
+    async function fetchMovies() {
+      const moviesResp = await fetch('/movies.json');
+      const movies = await (moviesResp.json());
+      movies.forEach((movie, idx) => { movie.id = idx + 1 });
+      setState({ movies });
+    }
+    fetchMovies();
+  }, []);
+
   return (
     <div className="App">
       <header className="App-header">
@@ -13,8 +25,8 @@ function App() {
       </header>
       <main className="MainContent">
         <Switch>
-          <Route path="/" component={MovieList} exact />
-          <Route path="/movie-detail/:id" component={MovieDetail} />
+          <Route path="/" exact render={(props) => <MovieList {...props} movies={state.movies} />} />
+          <Route path="/movie-detail/:id"  render={(props) => <MovieDetail {...props} movies={state.movies} />} />
         </Switch>      
       </main>
     </div>
