+++ clients/react-movies/src/pages/MovieDetail/MovieDetail.js
@@ -3,8 +3,13 @@ import { useParams } from 'react-router-dom';
 
 function MovieDetail (props) {
   const { id } = useParams();
+  const movie = props.movies.find(movie => movie.id === +id);  
+
   return (
-   <p>Movie Detail for movie with id {id}</p>
+    <div>
+      <h2>{movie.title}</h2>
+      <p>Movie Detail for movie with id {id}</p>
+    </div>
   );
 }
