+++ clients/react-movies/src/pages/MovieList/MovieList.css
@@ -7,7 +7,7 @@
   background-color: rgb(162, 150, 176);
   padding: 1em;
   margin: 1em;
-  border-radius: 4px;
+  border-radius: 2px;
 }
 
 .MovieCard a {
